==> notebooks/immunology/005_immuno_clocks_log_and_new_data/01_prepare_data.ipynb <==
# Debugging autoreload

%load_ext autoreload
%autoreload 2

# Load packages

import pickle
from scipy import stats
from glob import glob
import ast
import matplotlib.pyplot as plt
import seaborn as sns
import copy
import itertools
import numpy as np
import pandas as pd
import warnings
import pathlib
from tqdm import tqdm
import distinctipy
import matplotlib.patheffects as pe
import matplotlib.colors as mcolors
from statannotations.Annotator import Annotator
from scipy.stats import mannwhitneyu
import optuna
from matplotlib.colors import LinearSegmentedColormap
import matplotlib.cm
import matplotlib as mpl
from statsmodels.stats.multitest import multipletests
import re
from itertools import chain
from pathlib import Path
import os
import upsetplot
from sklearn.decomposition import PCA
from sklearn.random_projection import GaussianRandomProjection, SparseRandomProjection
from sklearn.manifold import MDS, Isomap, TSNE
import missingno as msno

# Prepare data

path = 'E:/YandexDisk/Work/bbd/immunology'
path_unn = f"E:/YandexDisk/Work/pydnameth/datasets/GPL21145/GSEUNN"

df = pd.read_excel(f"{path}/004_data_processing/data_GSEUNN_fmba_052025.xlsx", index_col=0)

feats = pd.read_excel(f"{path_unn}/data/immuno/feats_con.xlsx", index_col=0).index.to_list()
feats_log = [f"log({f})" for f in feats]

df = df.loc[
    (df['Status'] == 'Control') & (df['Region'].isin(['Central'])) | 
    (df['Group detailed'] == 'New 2025 UNN')
    ,
    ['Age', 'Sex', 'Region', 'Status', 'Group detailed', 'SImAge'] + feats_log
]

df.to_excel(f"{path}/005_immuno_clocks_log_and_new_data/data.xlsx", index_label='ID')
pd.DataFrame(index=feats_log).to_excel(f"{path}/005_immuno_clocks_log_and_new_data/feats.xlsx", index_label='Features')
